Added enemy attacks
base EnemyAttack and specific attack scripts for Gremlin, Mage, and Thief enemies

+++ PDT Ball Game/Assets/Scripts/PlayerHealth.cs
@@ -15,7 +15,7 @@ public class PlayerHealth : MonoBehaviour
         UpdateHealthDisplay();
     }
 
-    public void TakeDamage(int amount)
+    public void TakeDamage(float amount)
     {
         currentHealth -= amount;
         UpdateHealthDisplay();
